+++ packages/mobile/e2e/Screens/Settings/Options.spec.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
 const helpers = require('../../Helpers')
 import { by, device, element, expect } from 'detox'