+++ packages/mobile/src/ApplicationGroupContext.tsx
@@ -1,0 +1,4 @@
+import { ApplicationGroup } from '@Lib/ApplicationGroup'
+import React from 'react'
+
+export const ApplicationGroupContext = React.createContext<ApplicationGroup | undefined>(undefined)