+++ packages/mobile/src/Components/HeaderTitleView.tsx
@@ -9,14 +9,14 @@ type Props = {
 }
 
 const Container = styled.View`
-  /* background-color: ${props => props.theme.stylekitContrastBackgroundColor}; */
+  /* background-color: ${(props) => props.theme.stylekitContrastBackgroundColor}; */
   flex: 1;
   justify-content: center;
 
   ${Platform.OS === 'android' && 'align-items: flex-start; min-width: 100px;'}
 `
 const Title = styled.Text`
-  color: ${props => props.theme.stylekitForegroundColor};
+  color: ${(props) => props.theme.stylekitForegroundColor};
   font-weight: bold;
   font-size: 18px;
   text-align: center;
@@ -25,13 +25,13 @@ const SubTitle = styled.Text.attrs(() => ({
   adjustsFontSizeToFit: true,
   numberOfLines: 1,
 }))<{ color?: string }>`
-  color: ${props => props.color ?? props.theme.stylekitForegroundColor};
-  opacity: ${props => (props.color ? 1 : 0.6)};
+  color: ${(props) => props.color ?? props.theme.stylekitForegroundColor};
+  opacity: ${(props) => (props.color ? 1 : 0.6)};
   font-size: ${Platform.OS === 'android' ? 13 : 12}px;
   ${Platform.OS === 'ios' && 'text-align: center'}
 `
 
-export const HeaderTitleView: React.FC<Props> = props => (
+export const HeaderTitleView: React.FC<Props> = (props) => (
   <Container>
     <Title>{props.title}</Title>
     {props.subtitle && props.subtitle.length > 0 ? (