+++ packages/mobile/src/Components/TableSection.ts
@@ -3,5 +3,5 @@ import styled from 'styled-components/native'
 export const TableSection = styled.View`
   margin-top: 10px;
   margin-bottom: 10px;
-  background-color: ${props => props.theme.stylekitBackgroundColor};
+  background-color: ${(props) => props.theme.stylekitBackgroundColor};
 `