+++ packages/mobile/src/Screens/Notes/OfflineBanner.styled.ts
@@ -11,17 +11,17 @@ const Container = styled.View`
   padding: ${PADDING}px;
   border-width: 1px;
   border-radius: 4px;
-  border-color: ${props => props.theme.stylekitBorderColor};
+  border-color: ${(props) => props.theme.stylekitBorderColor};
 `
 const CenterContainer = styled.View`
   justify-content: center;
 `
 const UserIcon = styled(Icon)`
   font-size: 24px;
-  color: ${props => props.theme.stylekitInfoColor};
+  color: ${(props) => props.theme.stylekitInfoColor};
 `
 const ForwardIcon = styled(UserIcon)`
-  color: ${props => props.theme.stylekitNeutralColor};
+  color: ${(props) => props.theme.stylekitNeutralColor};
 `
 const TextContainer = styled.View`
   flex: 1;
@@ -30,12 +30,12 @@ const TextContainer = styled.View`
 const BoldText = styled.Text`
   font-size: 15px;
   font-weight: 600;
-  color: ${props => props.theme.stylekitForegroundColor};
+  color: ${(props) => props.theme.stylekitForegroundColor};
 `
 const SubText = styled.Text`
   margin-top: 2px;
   font-size: 11px;
-  color: ${props => props.theme.stylekitNeutralColor};
+  color: ${(props) => props.theme.stylekitNeutralColor};
 `
 
 export { Touchable, Container, CenterContainer, UserIcon, ForwardIcon, TextContainer, BoldText, SubText }