+++ packages/mobile/src/Screens/screens.ts
@@ -8,6 +8,7 @@ export const SCREEN_INPUT_MODAL_FILE_NAME = 'InputModalFileName'
 export const SCREEN_NOTE_HISTORY = 'NoteSessionHistory' as const
 export const SCREEN_NOTE_HISTORY_PREVIEW = 'NoteSessionHistoryPreview' as const
 export const SCREEN_UPLOADED_FILES_LIST = 'UploadedFilesList' as const
+export const SCREEN_INPUT_MODAL_WORKSPACE_NAME = 'InputModalWorkspaceName' as const
 
 export const SCREEN_SETTINGS = 'Settings'
 export const SCREEN_MANAGE_SESSIONS = 'ManageSessions' as const